+++ src/data/heroicons.js
@@ -7,7 +7,10 @@
 import arrowTrendingUp from "heroicons/24/outline/arrow-trending-up.svg?raw";
 import bars3 from "heroicons/24/outline/bars-3.svg?raw";
 import briefcase from "heroicons/24/outline/briefcase.svg?raw";
+import cake from "heroicons/24/outline/cake.svg?raw";
 import faceSmile from "heroicons/24/outline/face-smile.svg?raw";
+import language from "heroicons/24/outline/language.svg?raw";
+import mapPin from "heroicons/24/outline/map-pin.svg?raw";
 import sun from "heroicons/24/outline/sun.svg?raw";
 import xMark from "heroicons/24/outline/x-mark.svg?raw";
 
@@ -15,7 +18,10 @@ export const heroicons = {
   arrowTrendingUp,
   bars3,
   briefcase,
+  cake,
   faceSmile,
+  language,
+  mapPin,
   sun,
   xMark,
 };